feat(mailer): complete WELCOME email template
