+++ .changeset/lemon-shrimps-push.md
@@ -1,0 +1,5 @@
+---
+'@services/mailer': minor
+---
+
+complete the `WELCOME` email template (with i18n support)
