+++ services/mailer/package.json
@@ -16,7 +16,7 @@
     "dev": "vite dev --port 5007",
     "dev:wrangler": "wrangler pages dev .svelte-kit/cloudflare --port 5007",
     "build": "vite build --mode production",
-    "build:mjml": "mjml src/lib/server/templates/*.mjml -o src/lib/server/templates --config.mjmlConfigPath .mjmlconfig",
+    "build:mjml": "mjml src/lib/server/mjml/templates/*.template.mjml -o src/lib/server/mjml/templates --config.mjmlConfigPath .mjmlconfig --config.minify=true --config.beautify=false",
     "preview": "vite preview",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
