+++ services/mailer/src/lib/server/mjml/includes/header.mjml
@@ -1,0 +1,6 @@
+<mj-section>
+  <mj-column>
+    <mj-image width="40px" src="https://sveltevietnam.dev/logo/positive.svg" />
+    <mj-divider />
+  </mj-column>
+</mj-section>
