+++ services/mailer/src/lib/server/mjml/templates/welcome-en.template.mjml
@@ -1,0 +1,33 @@
+<mjml>
+  <mj-head>
+    <mj-title>{{title}}</mj-title>
+    <mj-include path="../includes/common-head.mjml" />
+  </mj-head>
+  <mj-body background-color="#ffffff">
+    <mj-include path="../includes/header.mjml" />
+
+    <mj-section padding-top="0">
+      <mj-column>
+        <mj-text padding-top="0">Hello {{name}},</mj-text>
+        <mj-text>
+          Thank you for visting
+          <a href="{{wwwURL}}">sveltevietnam.dev</a>.
+          We are happy to welcome you as a part of the Svelte Vietnam community.
+        </mj-text>
+        <mj-text>
+          If you have not already, join the official
+          <a href="https://discord.sveltevietnam.dev">Svelte Vietnam Discord</a>
+          to connect and receive more frequent updates from our community.
+        </mj-text>
+        <mj-text>
+          See you out there,
+        </mj-text>
+        <mj-text>
+          The Svelte Vietnam Team
+        </mj-text>
+      </mj-column>
+    </mj-section>
+
+    <mj-include path="../includes/no-reply-footer-en.mjml" />
+  </mj-body>
+</mjml>
